+++ src/components/AlignmentWithSiteBarPlot.jsx
@@ -4,18 +4,18 @@ const $ = require("jquery");
 const _ = require("underscore");
 const text_width = require("text-width");
 
-import fastaParser from "./fasta";
-import BaseAlignment from "./basealignment";
+import fastaParser from "./../helpers/fasta";
+import BaseAlignment from "./BaseAlignment.jsx";
 import SiteAxis from "./SiteAxis.jsx";
-import Placeholder from "./placeholder";
+import Placeholder from "./Placeholder.jsx";
 import SequenceAxis from "./SequenceAxis.jsx";
-import ScrollBroadcaster from "./scrollbroadcaster";
-import { nucleotide_color, nucleotide_text_color } from "./colors";
+import ScrollBroadcaster from "./../helpers/ScrollBroadcaster";
+import { nucleotide_color, nucleotide_text_color } from "./../helpers/colors";
 //import SiteBarPlot from "./SiteBarPlot.jsx";
 //toremove
 import TestBarPlot from "./TestBarPlot.jsx";
 
-require("./app.scss");
+//require("./app.scss");
 
 class AlignmentWithSiteBarPlot extends Component {
   constructor(props) {
