+++ src/components/BaseAlignment.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { nucleotide_color, nucleotide_text_color } from "./colors";
+import { nucleotide_color, nucleotide_text_color } from "./../helpers/colors";
 
 const d3 = require("d3");
 const _ = require("underscore");
