+++ src/components/ScaffoldViewer.jsx
@@ -4,16 +4,16 @@ const $ = require("jquery");
 const _ = require("underscore");
 const text_width = require("text-width");
 
-import fastaParser from "./fasta";
-import BaseAlignment from "./basealignment";
+import fastaParser from "./../helpers/fasta";
+import BaseAlignment from "./BaseAlignment.jsx";
 import SiteAxis from "./SiteAxis.jsx";
-import Placeholder from "./placeholder";
+import Placeholder from "./Placeholder.jsx";
 import SequenceAxis from "./SequenceAxis.jsx";
-import Scaffold from "./scaffold";
-import ScrollBroadcaster from "./scrollbroadcaster";
-import { nucleotide_color, nucleotide_text_color } from "./colors";
+import Scaffold from "./Scaffold.jsx";
+import ScrollBroadcaster from "./../helpers/ScrollBroadcaster";
+import { nucleotide_color, nucleotide_text_color } from "./../helpers/colors";
 
-require("./app.scss");
+//require("./app.scss");
 
 class ScaffoldViewer extends Component {
   constructor(props) {
