+++ src/components/SequenceAxis.jsx
