+++ src/components/navComponents/DropDownOfAlignmentsToView.jsx
