+++ src/helpers/fasta.js
